Add AddressHandler, OrderHandler, AdminMenu, UserMenu and users permissions
