+++ src/main/java/com/pranienawezwanie/orderservicesservice/handlers/UserHandler.java
@@ -2,15 +2,12 @@ package com.pranienawezwanie.orderservicesservice.handlers;
 
 import com.pranienawezwanie.orderservicesservice.database.AppUserDao;
 import com.pranienawezwanie.orderservicesservice.database.EntityDao;
-import com.pranienawezwanie.orderservicesservice.model.Address;
 import com.pranienawezwanie.orderservicesservice.model.AppUser;
 import com.pranienawezwanie.orderservicesservice.model.UserType;
 
 import java.util.Optional;
-import java.util.Scanner;
 
 public class UserHandler {
-    private Scanner scanner = new Scanner(System.in);
     private EntityDao<AppUser> appUserEntityDao = new EntityDao<>();
 
     public void handle(String[] words) {
@@ -22,28 +19,6 @@ public class UserHandler {
         } else if (words[1].equalsIgnoreCase("type") &&
                 words[2].equalsIgnoreCase("change")) {
             handleChangeUserType(words);
-        } else if (words[1].equalsIgnoreCase("addaddress")) {
-            handleAddAddress(words);
-        }
-    }
-
-    private void handleAddAddress(String[] words) {
-        Long id = Long.parseLong(words[2]);
-        Optional<AppUser> appUserOptional = appUserEntityDao.findById(AppUser.class, id);
-
-        if (appUserOptional.isPresent()) {
-            EntityDao<Address> addressEntityDao = new EntityDao<>();
-            AppUser appUser = appUserOptional.get();
-
-            Address address = Address.builder()
-                    .city(words[2])
-                    .houseNum(words[3])
-                    .postalCode(words[4])
-                    .street(words[5])
-                    .user(appUser)
-                    .build();
-
-            addressEntityDao.saveOrUpdate(address);
         }
     }
